move selected to capping
